Canvas revamp: wire extension, wire color picker The Canvas scene has been re-worked a little bit to include the save button previously part of the ChipDesigner and a color picker that allows the user to choose the color of the next new wire they make.
The color picker was an easy thing to add, and makes it easy to test the
wire drawing functionality but making it clear when you are making a new
wire and when you are extending an existing one. Improving the visual
representation of the wire may make this unnecessary, but the ability to
color code wires seems like a nice addition anyway.